fix: consolidate common logic between borrow modals
